Refactor model structure: remove model_objects, add models for products, offers, and discounts
