+++ python/models/discounts.py
@@ -1,0 +1,5 @@
+class Discount:
+    def __init__(self, product, description, discount_amount):
+        self.product = product
+        self.description = description
+        self.discount_amount = discount_amount
