+++ python/models/offers.py
@@ -1,0 +1,14 @@
+from enum import Enum
+
+
+class SpecialOfferType(Enum):
+    THREE_FOR_TWO = 1
+    TEN_PERCENT_DISCOUNT = 2
+    TWO_FOR_AMOUNT = 3
+    FIVE_FOR_AMOUNT = 4
+
+class Offer:
+    def __init__(self, offer_type, product, argument):
+        self.offer_type = offer_type
+        self.product = product
+        self.argument = argument
